+++ traits/src/lib.rs
@@ -18,10 +18,11 @@ pub use nft::{AccountIdOrCollectionNftTuple, Nft, NftInfo, RoyaltyInfo};
 pub use priority::Priority;
 pub use property::Property;
 pub use resource::{Resource, ResourceInfo};
+
 pub mod primitives {
 	pub type CollectionId = u32;
 	pub type ResourceId = u32;
-	pub type NftId = u32;
+	pub type NftId = u128;
 	pub type BaseId = u32;
 	pub type SlotId = u32;
 	pub type PartId = u32;
